+++ lib/src/constants.dart
@@ -1,7 +1,3 @@
-import 'dart:io';
-
-String defaultAssetFolder =
-    "lib${Platform.pathSeparator}assets${Platform.pathSeparator}";
 const String defaultAssetFile = "app_asset.dart";
 
 const String intro = """